+++ src/components/ControlPanel/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 
 import '../../style/ControlPanel.css';
-import BasemapSelector from '../BasemapSelector';
-import DebugDumpButton from '../DebugDumpButton';
-import RegionSelector from '../RegionSelector';
+import BasemapSelector from './BasemapSelector';
+import DebugDumpButton from './DebugDumpButton';
+import RegionSelector from './RegionSelector';
 
 declare const __ENVIRONMENT__: 'development' | 'production';
 if (__ENVIRONMENT__ === 'development') {
