+++ src/style/DebugDumpButton.css
@@ -2,4 +2,8 @@
   position: absolute;
   top: 0;
   right: 0;
+
+  background-color: #D04721;
+  color: #051729;
+  font-weight: bold;
 }
